cs329s_waymo_object_detection/model/model_serving/requirements.txt
opencv-python
numpy
requests
flask
pillow
torch
torchvision
google-cloud-storage
cs329s_waymo_object_detection
wandb